backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
networkx==3.4.2
nh3==0.2.18
slowapi==0.1.9
psycopg2-binary==2.9.9
aiosmtplib==3.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
